refactor: move corpus edit from Manager into singlCorpusEdit component.
